Initial commit for Light LFS management. Includes a git filter, which converts large files into hashes for storage and performs syncing to a specified local path. Also includes Automator services which should be used to mark and add (favorite) a file, thus also downloading it.
Notes: Sync operations are locally logged. Filter will be separated into two parts; (1) a hashing filter, and (2) a separate syncing system, to be called after favorites have been added. This syncing system will attempt a git commit, and then compare the local directory contents with a list of git-tracked files in order to determine what to download and what to delete.